+++ R/MeasureSurvCindex.R
@@ -5,21 +5,25 @@
 #' @template param_eps
 #'
 #' @description
-#' Calculates weighted concordance statistics, which, depending on the chosen weighting method
-#' and tied times solution, are equivalent to several proposed methods.
+#' Calculates weighted concordance statistics, which, depending on the chosen
+#' weighting method (`weight_meth`) and tied times parameter (`tiex`), are
+#' equivalent to several proposed methods.
+#' By default, no weighting is applied and this is equivalent to Harrell's C-index.
 #'
 #' @details
-#' For the Kaplan-Meier estimate of the training survival distribution, `S`, and
-#' the Kaplan-Meier estimate of the training censoring distribution, `G`:
+#' For the Kaplan-Meier estimate of the **training survival** distribution (\eqn{S}),
+#' and the Kaplan-Meier estimate of the **training censoring** distribution (\eqn{G}),
+#' we have the following options for time-independent concordance statistics
+#' (C-indexes) given the weighted method:
 #'
 #' `weight_meth`:
 #'
 #' - `"I"` = No weighting. (Harrell)
 #' - `"GH"` = Gonen and Heller's Concordance Index
-#' - `"G"` = Weights concordance by G^-1.
-#' - `"G2"` = Weights concordance by G^-2. (Uno et al.)
-#' - `"SG"` = Weights concordance by S/G (Shemper et al.)
-#' - `"S"` = Weights concordance by S (Peto and Peto)
+#' - `"G"` = Weights concordance by \eqn{1/G}.
+#' - `"G2"` = Weights concordance by \eqn{1/G^2}. (Uno et al.)
+#' - `"SG"` = Weights concordance by \eqn{S/G} (Shemper et al.)
+#' - `"S"` = Weights concordance by \eqn{S} (Peto and Peto)
 #'
 #' The last three require training data. `"GH"` is only applicable to [LearnerSurvCoxPH].
 #'
@@ -29,8 +33,11 @@
 #' computed on the same testing data.
 #'
 #' @section Parameter details:
-#' - `cutoff` (`numeric(1)`)\cr
-#' Cut-off time to evaluate concordance up to.
+#' - `t_max` (`numeric(1)`)\cr
+#' Cutoff time (i.e. time horizon) to evaluate concordance up to.
+#' - `p_max` (`numeric(1)`)\cr
+#' The proportion of censoring to evaluate concordance up to in the given dataset.
+#' When `t_max` is specified, this parameter is ignored.
 #' - `weight_meth` (`character(1)`)\cr
 #' Method for weighting concordance. Default `"I"` is Harrell's C. See details.
 #' - `tiex` (`numeric(1)`)\cr
@@ -44,6 +51,27 @@
 #' @template param_packages
 #' @template param_predict_type
 #' @template param_measure_properties
+#'
+#' @examples
+#' library(mlr3)
+#' task = tsk("rats")
+#' learner = lrn("surv.coxph")
+#' part = partition(task) # train/test split, stratified on `status` by default
+#' learner$train(task, part$train)
+#' p = learner$predict(task, part$test)
+#'
+#' # Harrell's C-index
+#' p$score(msr("surv.cindex")) # same as `p$score()`
+#'
+#' # Uno's C-index
+#' p$score(msr("surv.cindex", weight_meth = "G2"),
+#'         task = task, train_set = part$train)
+#'
+#' # Harrell's C-index evaluated up to a specific time horizon
+#' p$score(msr("surv.cindex", t_max = 97))
+#' # Harrell's C-index evaluated up to the time corresponding to 30% of censoring
+#' p$score(msr("surv.cindex", p_max = 0.3))
+#'
 #' @export
 MeasureSurvCindex = R6Class("MeasureSurvCindex",
   inherit = MeasureSurv,
@@ -51,7 +79,8 @@ MeasureSurvCindex = R6Class("MeasureSurvCindex",
     #' @description This is an abstract class that should not be constructed directly.
     initialize = function() {
       ps = ps(
-        cutoff = p_dbl(),
+        t_max = p_dbl(lower = 0),
+        p_max = p_dbl(0, 1),
         weight_meth = p_fct(levels = c("I", "G", "G2", "SG", "S", "GH"), default = "I"),
         tiex = p_dbl(0, 1, default = 0.5),
         eps = p_dbl(0, 1, default = 1e-3)
@@ -77,16 +106,34 @@ MeasureSurvCindex = R6Class("MeasureSurvCindex",
   private = list(
     .score = function(prediction, task, train_set, ...) {
       ps = self$param_set$values
+
+      # calculate t_max (cutoff time horizon)
+      if (is.null(ps$t_max) && !is.null(ps$p_max)) {
+        truth = prediction$truth
+        unique_times = unique(sort(truth[,"time"]))
+        surv = survival::survfit(truth ~ 1)
+        indx = which(1 - (surv$n.risk / surv$n) > ps$p_max)
+        if (length(indx) == 0) {
+          t_max = NULL # t_max calculated in `cindex()`
+        } else {
+          # first time point that surpasses the specified
+          # `p_max` proportion of censoring
+          t_max = surv$time[indx[1]]
+        }
+      } else {
+        t_max = ps$t_max
+      }
+
       if (ps$weight_meth == "GH") {
         return(gonen(prediction$crank, ps$tiex))
       } else if (ps$weight_meth == "I") {
-        return(cindex(prediction$truth, prediction$crank, ps$cutoff, ps$weight_meth, ps$tiex))
+        return(cindex(prediction$truth, prediction$crank, t_max, ps$weight_meth, ps$tiex))
       } else {
         if (is.null(task) | is.null(train_set)) {
-          stop("'task' and 'train_set' required for all weighted c-index (except GH).")
+          stop("'task' and 'train_set' required for all weighted C-indexes (except GH).")
         }
-        return(cindex(prediction$truth, prediction$crank, ps$cutoff, ps$weight_meth,
-          ps$tiex, task$truth(train_set), ps$eps))
+        return(cindex(prediction$truth, prediction$crank, t_max, ps$weight_meth,
+                      ps$tiex, task$truth(train_set), ps$eps))
       }
     }
   )